clean poc
