+++ OPlan/package.json
@@ -10,9 +10,10 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "@mui/material": "^6.4.3",
+    "opml": "^0.5.7",
     "react": "^18.3.1",
-    "react-dom": "^18.3.1",
-    "opml": "^0.5.7"
+    "react-dom": "^18.3.1"
   },
   "devDependencies": {
     "@eslint/js": "^9.17.0",
